Funciona toda la logica de torneos y tenista

+++ src/main/kotlin/com/example/torneo_kotlin/Rest/torneo/services/TorneoService.kt
@@ -1,5 +1,10 @@
 package com.example.torneo_kotlin.Rest.torneo.services
 
+import com.example.torneo_kotlin.Rest.tenistas.dto.TenistaDTO
+import com.example.torneo_kotlin.Rest.tenistas.exceptions.TenistaNoEncontradoException
+import com.example.torneo_kotlin.Rest.tenistas.mappers.toDTO
+import com.example.torneo_kotlin.Rest.tenistas.models.Tenista
+import com.example.torneo_kotlin.Rest.tenistas.repositories.TenistaRepositorio
 import com.example.torneo_kotlin.Rest.torneo.dto.CrearTorneoDTO
 import com.example.torneo_kotlin.Rest.torneo.dto.TorneoDTO
 import com.example.torneo_kotlin.Rest.torneo.dto.UpdateTorneoDTO
@@ -7,23 +12,29 @@ import com.example.torneo_kotlin.Rest.torneo.dto.UpdateTorneoParcialDTO
 import com.example.torneo_kotlin.Rest.torneo.exceptions.TorneoNoEncontrado
 import com.example.torneo_kotlin.Rest.torneo.mappers.toDTO
 import com.example.torneo_kotlin.Rest.torneo.mappers.toEntity
-import com.example.torneo_kotlin.Rest.torneo.models.Categoria
-import com.example.torneo_kotlin.Rest.torneo.models.Tipos
+import com.example.torneo_kotlin.Rest.torneo.models.Torneos
 import com.example.torneo_kotlin.Rest.torneo.repositories.TorneoRepositorio
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
+
 @Service
 @Transactional
-class TorneoService(private val torneoRepositorio: TorneoRepositorio) {
+class TorneoService(
+    private val torneoRepositorio: TorneoRepositorio,
+    private val tenistaRepositorio: TenistaRepositorio,
+) {
+
 
     fun obtenerTorneos(): List<TorneoDTO> {
         return torneoRepositorio.findAll().map { it.toDTO() }
     }
 
 
-    fun createTorneo(dto: CrearTorneoDTO): TorneoDTO =
-        torneoRepositorio.save(dto.toEntity()).toDTO()
+    fun createTorneo(dto: CrearTorneoDTO): TorneoDTO {
+        val torneoCreado = torneoRepositorio.save(dto.toEntity()).toDTO()
+        return torneoCreado
+    }
 
     fun optenerTorneoPorId(id: Long): TorneoDTO {
         val torneo =
@@ -47,24 +58,39 @@ class TorneoService(private val torneoRepositorio: TorneoRepositorio) {
 
     fun deleteTorneo(id: Long) {
         if (!torneoRepositorio.existsById(id)) {
-            throw RuntimeException("Torneo no encontrado!")
+            throw TorneoNoEncontrado("Torneo no encontrado!")
         }
+
+        if (tenistaRepositorio.existsByTorneoId(id)) {
+            throw TorneoNoEncontrado("No se puede eliminar el torneo porque está siendo utilizado por uno o más tenistas.")
+        }
+
         torneoRepositorio.deleteById(id)
     }
 
-
-    fun validarCrearTorneoDTO(dto: CrearTorneoDTO) {
-        // Valida que la modalidad está dentro de los valores permitidos
-        if (dto.modalidad !in Tipos.values()) {
-            throw IllegalArgumentException("La modalidad especificada no es válida.")
+    @Throws(Exception::class)
+    fun inscribirTenistaEnTorneo(torneoId: Long, tenistaId: Long): TenistaDTO {
+        val torneo: Torneos = torneoRepositorio.findById(torneoId).orElseThrow {
+            TorneoNoEncontrado("Torneo no encontrado")
         }
 
-        // Valida que la categoría está dentro de los valores permitidos
-        if (dto.categoria !in Categoria.values()) {
-            throw IllegalArgumentException("La categoría especificada no es válida.")
+        val tenista: Tenista = tenistaRepositorio.findById(tenistaId).orElseThrow {
+            TenistaNoEncontradoException("Tenista no encontrado")
+        }
+        if (tenista.torneo?.id == torneoId) {
+            throw TorneoNoEncontrado("El tenista ya está inscrito en el torneo")
+        }
+        if (torneo.vacantes > 0) {
+            torneo.vacantes -= 1
+            tenista.torneo = torneo
+            torneoRepositorio.save(torneo)
+            tenistaRepositorio.save(tenista)
+            return tenista.toDTO()
+        } else {
+            throw TorneoNoEncontrado("No hay vacantes disponibles")
         }
 
-        // Otros campos pueden ser validados según sea necesario
     }
+
 
 }
